Changes requested - Make button reuseable, make code for matting more user friendly
