+++ src/Pages/OpenEmailPage/index.jsx
@@ -1,18 +1,45 @@
-import { BrowserRouter, Route, Routes, Router } from "react-router-dom"
-import OpenEmails from "../../Components/OpenEmails"
-import { useEffect } from "react"
+import { useEffect, useState } from "react"
+import { useParams } from "react-router-dom"
+import GreyButton from "../../Components/GreyButton"
 
+function OpenEmails(){
+    const { id } = useParams()
+    const [emailId, setEmailId] = useState ("")
+    const [emailSubject, setEmailSubject] = useState ("")
+    const [emailBody, setEmailBody] = useState ("")
+    const [emailFrom, setEmailFrom] = useState ("")
+    const [emailFromName, setEmailFromName] = useState ("")
+    const [emailDate, setEmailDate] = useState ("")
 
-function OpenEmailPage(){
+    useEffect (() => {
+        fetch(`https://email-client-api.dev.io-academy.uk/emails/${id}`)
+        .then(res => res.json())
+        .then(data => {
+            const email = data.data.email
+            setEmailId(email.id)
+            setEmailSubject(email.subject)
+            setEmailBody(email.body)
+            setEmailFrom(email.email)
+            setEmailFromName(email.name)
+            setEmailDate(email.date_created)
+        })
+    }, [id])
 
-    return(
-        <BrowserRouter>
-            <Router>
-                <Routes>
-                    <Route path="/emails/:id" element={<OpenEmails />} />
-                </Routes>
-            </Router>
-        </BrowserRouter>
+    return (
+        <main className="my-6 border-y-4 border-slate-100 px-6">
+            <div className="grid justify-items-end pt-4"> 
+                <GreyButton link="/pages/HomePage" text="Return to Inbox"> 
+                </GreyButton> 
+            </div>
+            <div className="pt-5 font-bold 
+            flex flex-row justify-between">
+                <h4>{emailFromName}</h4>
+                <p>{emailDate}</p>
+            </div>
+            <h5 className="py-2 font-bold">{emailFrom}</h5>
+            <h3 className="text-2xl font-semibold">{emailSubject}</h3>
+            <p className="mb-4 pt-10 flex flex-row justify-start">{emailBody}</p>
+        </main>
     )
 }
-export default OpenEmailPage
+export default OpenEmails
